Produces plots of CALCOFI data

This function produces plots of CALCOFI-originated time series data.

calcofi_plot(
  csv,
  x_fld = "year",
  y_fld = "avg_larvae_count_per_volume_sampled",
  y_trans = "log(y + 1)",
  x_lab = "Year",
  y_lab = "ln(mean abundance + 1)",
  title = NULL,
  yrs_recent = 5,
  interactive = T,
  in_loop = F
)

Arguments

csv
The CALCOFI time series data, in csv format, to be plotted.

x_fld
The column of the time series data to be used for the x-axis of the plot.

y_fld
The column of the time series data to be used for the y-axis of the plot.

y_trans
The transformation to be performed on the data used for the y-axis.

x_lab
The label for the x-axis on the plot.

y_lab
The label for the y-axis on the plot.

title
The label for the title of the plot.

yrs_recent
The number of most recent years to be shaded in the plot.

interactive
A Boolean variable indicating whether the plot is to be interactive or not.

in_loop
A Boolean variable indicating whether an error condition exists.

Value

The output is a plot of time series data.

Examples

if (FALSE) {
calcofi_plot(csv = "https://raw.githubusercontent.com/marinebon/calcofi-analysis/master/data/Anchovy_CINMS.csv", title = "Anchovy - CINMS Region")
}