Generate a plot of intertidal monitoring data

This function generates a plot of time series data collected by the Rocky Intertidal Monitoring program.

plot_intertidal_nms(
  d_csv,
  NMS,
  spp,
  sp_name,
  nms_rgns,
  spp_targets = NULL,
  fld_val = "pct_cover",
  label_y = "Annual Mean Percent Cover (%)",
  label_x = "Year",
  nms_skip_regions = c("OCNMS", "MBNMS")
)

Arguments

d_csv
A csv file containing the time series data.

NMS
The National Marine Sanctuary code.

spp
BEN

sp_name
BEN

nms_rgns
BEN

spp_targets
BEN

fld_val
The column of data to provide the y-axis to be plotted.

label_y
The label for the y-axis for the plot.

label_x
The label for the x-axis for the plot.

nms_skip_regions
National Marine Sanctuaries to be skipped (using sanctuary codes).

Value

This function returns a dygraph object of the plotted time series data.

Examples

nms <- "cinms"
d_csv <- "/Volumes/GoogleDrive/Shared drives/NMS/data/github_info-intertidal_data/sanctuary_species_percentcover.csv"
nms_rgns_csv    <- file.path(dir_pfx, "MARINe_graphs.xlsx - sites in regions.csv")

#> Error in file.path(dir_pfx, "MARINe_graphs.xlsx - sites in regions.csv"): object 'dir_pfx' not found

nms_rgns1 <- readr::read_csv(nms_rgns_csv) %>%
  tidyr::fill(nms) %>%
  dplyr::group_by(nms) %>%
  tidyr::fill(region) %>%
  dplyr::mutate(rgn = region)

#> Error in is.connection(x): object 'nms_rgns_csv' not found

plot_intertidal_nms(d_csv, "CINMS", "MYTCAL", "California Mussels", nms_rgns1)

#> 
#> ── Column specification ────────────────────────────────────────────────────────
#> cols(
#>   site = col_character(),
#>   date = col_date(format = ""),
#>   pct_cover = col_double(),
#>   nms = col_character(),
#>   sp = col_character(),
#>   sp_target = col_character()
#> )

#> Error in is.data.frame(y): object 'nms_rgns1' not found